Refactor to clean up cli.hpp significantly.
