+++ src/mlpack/core/util/cli_impl.hpp
@@ -224,185 +224,6 @@ typename util::ParameterType<T>::type& CLI::GetUnmappedParam(
   return *boost::any_cast<typename util::ParameterType<T>::type>(&d.value);
 }
 
-namespace util {
-
-//! This overload is called when nothing special needs to happen to the name of
-//! the parameter.
-template<typename T>
-std::string MapParameterName(
-    const std::string& identifier,
-    const typename boost::disable_if<arma::is_arma_type<T>>::type* /* junk */,
-    const typename boost::disable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  return identifier;
-}
-
-//! This overload is called when T == ParameterType<T>::value.  An overload for
-//! matrices is in cli.cpp.
-template<typename T>
-T& HandleParameter(
-    typename ParameterType<T>::type& value,
-    ParamData& /* d */,
-    const typename boost::disable_if<arma::is_arma_type<T>>::type* /* junk */,
-    const typename boost::disable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  return value;
-}
-
-//! This is called for matrices, which have a different boost name.
-template<typename T>
-std::string MapParameterName(
-    const std::string& identifier,
-    const typename boost::enable_if<arma::is_arma_type<T>>::type* /* junk */)
-{
-  return identifier + "_file";
-}
-
-//! This overload is called for matrices, which return a different type.
-template<typename T>
-T& HandleParameter(
-    typename ParameterType<T>::type& value,
-    ParamData& d,
-    const typename boost::enable_if<arma::is_arma_type<T>>::type* /* junk */)
-{
-  // If the matrix is an input matrix, we have to load the matrix.  'value'
-  // contains the filename.  It's possible we could load empty matrices many
-  // times, but I am not bothered by that---it shouldn't be something that
-  // happens.
-  T& matrix = *boost::any_cast<T>(&d.mappedValue);
-  if (d.input && !d.loaded)
-  {
-    data::Load(value, matrix, true, !d.noTranspose);
-    d.loaded = true;
-  }
-
-  return matrix;
-}
-
-//! This is called for serializable mlpack objects, which have a different boost
-//! name.
-template<typename T>
-std::string MapParameterName(
-    const std::string& identifier,
-    const typename boost::enable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  return identifier + "_file";
-}
-
-//! This is called for serializable mlpack objects, which have a different boost
-//! name.
-template<typename T>
-T& HandleParameter(
-    typename ParameterType<T>::type& value,
-    ParamData& d,
-    const typename boost::enable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  // If the model is an input model, we have to load it from file.  'value'
-  // contains the filename.
-  T& model = *boost::any_cast<T>(&d.mappedValue);
-  if (d.input && !d.loaded)
-  {
-    data::Load(value, "model", model, true);
-  }
-
-  return model;
-}
-
-//! Print an option.
-template<typename T>
-void PrintParamImpl(
-    const ParamData& data,
-    const typename boost::disable_if<arma::is_arma_type<T>>::type* /* junk */,
-    const typename boost::disable_if<IsStdVector<T>>::type* /* junk */,
-    const typename boost::disable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  Log::Info << boost::any_cast<T>(data.value);
-}
-
-//! Print a vector option.
-template<typename T>
-void PrintParamImpl(
-    const ParamData& data,
-    const typename boost::enable_if<IsStdVector<T>>::type* /* junk */)
-{
-  const T& t = boost::any_cast<T>(data.value);
-
-  for (size_t i = 0; i < t.size(); ++i)
-    Log::Info << t[i] << " ";
-}
-
-//! Print a matrix option (this just prints the filename).
-template<typename T>
-void PrintParamImpl(
-    const ParamData& data,
-    const typename boost::enable_if<arma::is_arma_type<T>>::type* /* junk */)
-{
-  Log::Info << boost::any_cast<std::string>(data.value);
-}
-
-//! Print a model option (this just prints the filename).
-template<typename T>
-void PrintParamImpl(
-    const ParamData& data,
-    const typename boost::enable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  Log::Info << boost::any_cast<std::string>(data.value);
-}
-
-//! Output an option.
-template<typename T>
-void OutputParamImpl(
-    const ParamData& data,
-    const typename boost::disable_if<arma::is_arma_type<T>>::type* /* junk */,
-    const typename boost::disable_if<IsStdVector<T>>::type* /* junk */,
-    const typename boost::disable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  std::cout << data.name << ": " << *boost::any_cast<T>(&data.value);
-}
-
-//! Output a vector option.
-template<typename T>
-void OutputParamImpl(
-    const ParamData& data,
-    const typename boost::enable_if<IsStdVector<T>>::type* /* junk */)
-{
-  std::cout << data.name << ": ";
-  const T& t = *boost::any_cast<T>(&data.value);
-  for (size_t i = 0; i < t.size(); ++i)
-    std::cout << t[i] << " ";
-}
-
-//! Output a matrix option (this saves it to file).
-template<typename T>
-void OutputParamImpl(
-    const ParamData& data,
-    const typename boost::enable_if<arma::is_arma_type<T>>::type* /* junk */)
-{
-  const T& output = *boost::any_cast<T>(&data.mappedValue);
-  const std::string& filename = *boost::any_cast<std::string>(&data.value);
-
-  if (output.n_elem > 0 && filename != "")
-    data::Save(filename, output, false, !data.noTranspose);
-}
-
-//! Output a model option (this saves it to file).
-template<typename T>
-void OutputParamImpl(
-    const ParamData& data,
-    const typename boost::enable_if<data::HasSerialize<T>>::type* /* junk */)
-{
-  // The const cast is necessary here because Serialize() can't ever be marked
-  // const.  In this case we can assume it though, since we will be saving and
-  // not loading.
-  T& output = const_cast<T&>(*boost::any_cast<T>(&data.mappedValue));
-  const std::string& filename = *boost::any_cast<std::string>(&data.value);
-
-  if (filename != "")
-    data::Save(filename, "model", output);
-}
-
-} // namespace util
-
 template<typename T>
 void CLI::AddOption(
     const char* optId,
